Update build configuration and documentation (#7)
- Added build_log.txt to .gitignore to prevent unnecessary log files from being tracked.
- Updated BUILD.md to include new toolchain notes and build instructions for C23 compliance.
- Enhanced CMakeLists.txt to enforce C23 standards and added options for stricter warning settings.
- Improved README.md with new standards section detailing C23 target and internal code preferences.
- Made minor adjustments to various source files to suppress unused parameter warnings and improve code clarity.

+++ src/server/sv_snapshot.c
@@ -667,7 +667,7 @@ static void SV_BuildClientSnapshot( client_t *client ) {
 
 	// now that all viewpoint's areabits have been OR'd together, invert
 	// all of them to make it a mask vector, which is what the renderer wants
-	for ( i = 0; i < MAX_MAP_AREA_BYTES/sizeof(int); i++ ) {
+	for ( i = 0; (size_t) i < MAX_MAP_AREA_BYTES/sizeof(int); i++ ) {
 		((int *)frame->areabits)[i] = ((int *)frame->areabits)[i] ^ -1;
 	}
